+++ docs/changelog.md
@@ -1,7 +1,8 @@
 This page tracks releases of `earthmover`, with a summary of what was changed, fixed, added in each new version.
 
-### Unreleased changes
+## Unreleased changes
 
+* feature: Add `melt` and `pivot` as dataframe operations
 * feature: Introduce precedence order for resolving configuration when using project composition
 
 ## 2025 releases